move test image helper into internal package
