+++ test/util.go
@@ -3,9 +3,7 @@ package test
 import (
 	"fmt"
 	"image"
-	"image/color"
 	"image/png"
-	"math"
 	"os"
 	"path/filepath"
 	"testing"
@@ -128,23 +126,6 @@ func MoveMouse(c fyne.Canvas, pos fyne.Position) {
 	if tc != nil {
 		tc.hovered = hovered
 	}
-}
-
-// NewCheckedImage returns a new black/white checked image with the specified size
-// and the specified amount of horizontal and vertical tiles.
-func NewCheckedImage(w, h, hTiles, vTiles int) image.Image {
-	img := image.NewNRGBA(image.Rect(0, 0, w, h))
-	colors := []color.Color{color.White, color.Black}
-	tileWidth := float64(w) / float64(hTiles)
-	tileHeight := float64(h) / float64(vTiles)
-	for y := 0; y < h; y++ {
-		yTile := int(math.Floor(float64(y) / tileHeight))
-		for x := 0; x < w; x++ {
-			xTile := int(math.Floor(float64(x) / tileWidth))
-			img.Set(x, y, colors[(xTile+yTile)%2])
-		}
-	}
-	return img
 }
 
 // Scroll scrolls at an absolute position on the canvas.
