+++ test/util_test.go
@@ -1,7 +1,6 @@
 package test_test
 
 import (
-	"fmt"
 	"image"
 	"image/color"
 	"image/draw"
@@ -94,27 +93,6 @@ func TestDrag(t *testing.T) {
 		DraggedY:   42,
 	}, d.event)
 	assert.True(t, d.wasDragged)
-}
-
-func TestNewCheckedImage(t *testing.T) {
-	img := test.NewCheckedImage(10, 10, 5, 2)
-	expectedColorValues := [][]uint8{
-		{0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff},
-		{0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff},
-		{0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff},
-		{0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff},
-		{0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff},
-		{0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0},
-		{0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0},
-		{0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0},
-		{0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0},
-		{0x0, 0x0, 0xff, 0xff, 0x0, 0x0, 0xff, 0xff, 0x0, 0x0},
-	}
-	for y, xv := range expectedColorValues {
-		for x, v := range xv {
-			assert.Equal(t, color.NRGBA{R: v, G: v, B: v, A: 0xff}, img.At(x, y), fmt.Sprintf("color value at %d,%d", x, y))
-		}
-	}
 }
 
 func TestScroll(t *testing.T) {
